Feature/order logic (#2)
* refactor : Position -> TradeDirection 변경 및, Status,Type,Action 추가

* feat : 조건부 주문 추가

* feat : 각종 enum 추가 및 조건부 주문 추가

* refactor : leverage type 변경

* refactor : OrderService -> TradingService 직관적인 네이밍 변경 및 SimulatedTradingService 로직 추가

* feat : TODO 추가 및 OrderStatus Update 로직 추가

* refactor : OrderService -> TradingService 직관적인 네이밍 변경

* refactor : OrderService -> TradingService 직관적인 네이밍 변경

* refactor : TODO 작성 및 OrderService -> TradingService 직관적인 네이밍 변경

* feat : findOpenOrderByBotIdAndSymbol - OrderStatus가 None이며 PARTIALLY_FILLED. 즉, OPEN되어있는Order 가져오는 것 추가

* feat : OrderService 추가

+++ src/main/java/com/quant/craft/ordermanagement/dto/OrderDto.java
@@ -24,7 +24,7 @@ public class OrderDto {
     private BigDecimal stop;
     private BigDecimal sl;
     private BigDecimal tp;
-    private BigDecimal leverage;
+    private int leverage;
     private Boolean exclusiveOrders;
     private Boolean hedgeMode;
